test : 메소드 이동에 따른 테스트 이동
