+++ src/main/java/domain/blackjack/BlackJackResult.java
@@ -3,8 +3,6 @@ package domain.blackjack;
 import domain.participant.Participant;
 
 import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Set;
 
 public class BlackJackResult {
 
@@ -19,10 +17,6 @@ public class BlackJackResult {
                 .filter(isWin -> WinStatus.LOSE == isWin)
                 .count();
         return (int) dealerWinCount;
-    }
-
-    public Set<Map.Entry<Participant, WinStatus>> getEntry() {
-        return resultByParticipant.entrySet();
     }
 
     public int getTotalCount() {
